Modify FindUsages callback to accept locations
The OmniSharp#actions#usages#Find() function previously accepted an
option callback function, and when it was present the callback was
called with the number of usages, *after* the usages were sent to the
configured g:OmniSharp_selector_findusages tool or quickfix list.

This change now sends the full locations list to the callback when it
exists, *instead of* using a selector or quickfix list.

This makes the function more easily scriptable.